requirements.txt
langchain==0.0.335
openai==1.2.3
python-decouple==3.8
streamlit==1.28.1